requirements.txt
pytest==7.2.1
requests==2.28.2
allure-pytest==2.12
